+++ src/main/java/dk/sdu/se_f22/productmodule/infrastructure/domain/ProductInfIndexImpl.java
@@ -4,8 +4,8 @@ import dk.sdu.se_f22.brandmodule.stemming.Stemmer;
 import dk.sdu.se_f22.productmodule.infrastructure.ProductIndexInfrastructure;
 import dk.sdu.se_f22.productmodule.infrastructure.data.TokenParameter;
 import dk.sdu.se_f22.productmodule.irregularwords.Data.IrregularWords;
-import dk.sdu.se_f22.productmodule.management.ProductAttribute;
-import dk.sdu.se_f22.productmodule.management.BaseProduct;
+import dk.sdu.se_f22.productmodule.management.domain_persistance.ProductAttribute;
+import dk.sdu.se_f22.productmodule.management.domain_persistance.BaseProduct;
 
 import java.util.ArrayList;
 import java.util.List;
