+++ src/main/java/dk/sdu/se_f22/productmodule/infrastructure/domain/ProductInfSearchImpl.java
@@ -1,7 +1,7 @@
 package dk.sdu.se_f22.productmodule.infrastructure.domain;
 
 import dk.sdu.se_f22.searchmodule.infrastructure.interfaces.IndexingModule;
-import dk.sdu.se_f22.productmodule.management.BaseProduct;
+import dk.sdu.se_f22.productmodule.management.domain_persistance.BaseProduct;
 
 import java.util.ArrayList;
 import java.util.List;
