+++ src/main/java/dk/sdu/se_f22/productmodule/management/domain_persistance/BaseProduct.java
@@ -1,4 +1,4 @@
-package dk.sdu.se_f22.productmodule.management;
+package dk.sdu.se_f22.productmodule.management.domain_persistance;
 
 import org.jetbrains.annotations.Nullable;
