+++ src/main/java/dk/sdu/se_f22/productmodule/management/domain_persistance/ProductManager.java
@@ -1,4 +1,4 @@
-package dk.sdu.se_f22.productmodule.management;
+package dk.sdu.se_f22.productmodule.management.domain_persistance;
 
 import dk.sdu.se_f22.productmodule.infrastructure.ProductIndexInfrastructure;
